Add Version 5.0

+++ src/Kernel/Setup.cs
@@ -27,21 +27,12 @@ namespace ProPharmacyManager.Kernel
         {
             InitializeComponent();
         }
-        
+
         private void Setup_Load(object sender, EventArgs e)
         {
-            MD5.Create(serial.Text);
             if (this.Text == "تنصيب البرنامج")
             {
                 Install.Text = "اعداد";
-            }
-            else
-            {
-                serial.Visible = false;
-                label8.Visible = false;
-                Exit.Location = new System.Drawing.Point(41, 318);
-                Install.Location = new System.Drawing.Point(277, 318);
-                this.Size = new System.Drawing.Size(495, 429);
             }
             try
             {
@@ -60,7 +51,7 @@ namespace ProPharmacyManager.Kernel
             }
         }
 
-        private void button2_Click(object sender, EventArgs e)
+        private void Exit_Click(object sender, EventArgs e)
         {
             if (this.Text == "تنصيب البرنامج")
             {
@@ -71,20 +62,13 @@ namespace ProPharmacyManager.Kernel
                 this.Close();
             }
         }
-        private void button1_Click(object sender, EventArgs e)
+
+        private void Install_Click(object sender, EventArgs e)
         {
             try
             {
                 IniFile file = new IniFile(Constants.SetupConfigPath);
-                if (serial.Text == "" && this.Text == "تنصيب البرنامج")
-                {
-                    MessageBox.Show("انت لم تقوم بادخال الحروف المتسلسله");
-                }
-                else if (serial.Text != "9499-8790-0047-1128" && this.Text == "تنصيب البرنامج")
-                {
-                    MessageBox.Show("السريل خاطئ");
-                }
-                else if (serial.Text == "9499-8790-0047-1128" && this.Text == "تنصيب البرنامج")
+                if (this.Text == "تنصيب البرنامج")
                 {
                     if (!File.Exists(Constants.SetupConfigPath))
                     {
@@ -95,25 +79,30 @@ namespace ProPharmacyManager.Kernel
                             Directory.CreateDirectory(Application.StartupPath + @"\BackUp\");
                         }
                         DataHolder.CreateConnection(file.ReadString("MySql", "Username"), file.ReadString("MySql", "Password"), file.ReadString("MySql", "Host"));
-                        CreateDB.Createdb(DBName.Text, DBName.Text);
+                        CreateDB.Createdb(DBName.Text);
                         DataHolder.CreateConnection(file.ReadString("MySql", "Username"), file.ReadString("MySql", "Password"), file.ReadString("MySql", "Database"), file.ReadString("MySql", "Host"));
                         CreateDB.CreateTables();
-                        MessageBox.Show("تم تنصيب الاعدادات");
+                        MessageBox.Show("تم تنصيب الاعدادات\nمن فضلك انشاء حساب جديد لتتمكن من الدخول");
+                        Register reg = new Register();
+                        reg.ShowDialog();
                         this.Close();
                     }
                 }
-                else if (File.Exists(Constants.SetupConfigPath))
+                else if (this.Text == "اعدادات البرنامج")
                 {
-                    BackUp.NewDbBackup();
-                    DataHolder.CreateConnection(file.ReadString("MySql", "Username"), file.ReadString("MySql", "Password"), file.ReadString("MySql", "Host"));
-                    CreateDB.Createdb(file.ReadString("MySql", "Database"), DBName.Text);
-                    File.Delete(Constants.SetupConfigPath);
-                    string[] lines = { "[MySql]", "Host=" + DBHost.Text, "Username=" + DBUser.Text, "Password=" + DBPass.Text, "Database=" + DBName.Text };
-                    File.WriteAllLines(Constants.SetupConfigPath, lines);
-                    DataHolder.CreateConnection(file.ReadString("MySql", "Username"), file.ReadString("MySql", "Password"), file.ReadString("MySql", "Database"), file.ReadString("MySql", "Host"));
-                    BackUp.NewDbRestore();
-                    BillsTable.LBN();
-                    MessageBox.Show("تم تغيير الاعدادات");
+                    if (File.Exists(Constants.SetupConfigPath))
+                    {
+                        BackUp.NewDbBackup();
+                        DataHolder.CreateConnection(file.ReadString("MySql", "Username"), file.ReadString("MySql", "Password"), file.ReadString("MySql", "Host"));
+                        CreateDB.Createdb(file.ReadString("MySql", "Database"), DBName.Text);
+                        File.Delete(Constants.SetupConfigPath);
+                        string[] lines = { "[MySql]", "Host=" + DBHost.Text, "Username=" + DBUser.Text, "Password=" + DBPass.Text, "Database=" + DBName.Text };
+                        File.WriteAllLines(Constants.SetupConfigPath, lines);
+                        DataHolder.CreateConnection(file.ReadString("MySql", "Username"), file.ReadString("MySql", "Password"), file.ReadString("MySql", "Database"), file.ReadString("MySql", "Host"));
+                        BackUp.NewDbRestore();
+                        BillsTable.LBN();
+                        MessageBox.Show("تم تغيير الاعدادات");
+                    }
                 }
             }
             catch (Exception ee)
@@ -121,5 +110,6 @@ namespace ProPharmacyManager.Kernel
                 Program.SaveException(ee);
             }
         }
+
     }
 }
